# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/invoices": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Generates a VAT invoice (5% UAE VAT) for a deal. Invoice number is auto-assigned as INV-YYYY-NNNN.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Create VAT invoice",
                "parameters": [
                    {
                        "description": "Invoice payload",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "deal_id": {
                                    "type": "string"
                                },
                                "subtotal": {
                                    "type": "number"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.VATInvoice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Get invoice",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.VATInvoice"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}/send": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Transitions the invoice status from draft → sent.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Mark invoice as sent",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}/status": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Manually set invoice status to draft|sent|paid.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoice status",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Invoice UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "New status",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "object",
                            "properties": {
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "id": {
                                    "type": "string"
                                },
                                "status": {
                                    "type": "string"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "properties": {
                                "error": {
                                    "type": "string"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.InvoiceStatus": {
            "type": "string",
            "enum": [
                "draft",
                "sent",
                "paid"
            ],
            "x-enum-varnames": [
                "InvoiceDraft",
                "InvoiceSent",
                "InvoicePaid"
            ]
        },
...
        "github_com_dynamicweblab_masaar-crm_internal_domain.VATInvoice": {
            "type": "object",
            "properties": {
                "deal_id": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "invoice_no": {
                    "type": "string"
                },
                "issued_at": {
                    "type": "string"
                },
                "qr_payload": {
                    "type": "string"
                },
                "status": {
                    "$ref": "#/definitions/github_com_dynamicweblab_masaar-crm_internal_domain.InvoiceStatus"
                },
                "subtotal": {
                    "type": "number"
                },
                "total": {
                    "type": "number"
                },
                "vat_amount": {
                    "type": "number"
                },
                "vat_rate": {
                    "type": "number"
                }
            }
        },
...
    }
}